type refactoring

+++ 01_MyDevServer/00_TB_IOCP_CHAT/ChatApp/LogicMain.h
@@ -49,6 +49,6 @@ private:
 	unsigned int mMaxRoomCnt = 0;
 
 	using RECV_PKT_TYPE = void (LogicMain::*)(unsigned int,unsigned int, char*);
-	std::unordered_map<int, RECV_PKT_TYPE> mRecvFuncDic;
+	std::unordered_map<unsigned short, RECV_PKT_TYPE> mRecvFuncDic;
 };
